requirements.txt
# openai==0.28.0
openai== 1.37.1
python-dotenv==1.0.1
streamlit==1.37.0